+++ Marlin/src/inc/Conditionals_LCD.h
@@ -36,6 +36,21 @@
   #define FYSETC_MINI_12864_2_1
 #endif
 
+// Updated DGUS_UI shorthand single option can be used, or old settings, for now
+#if DGUS_UI_IS(ORIGIN)
+  #define DGUS_LCD_UI_ORIGIN
+#elif DGUS_UI_IS(FYSETC)
+  #define DGUS_LCD_UI_FYSETC
+#elif DGUS_UI_IS(HIPRECY)
+  #define DGUS_LCD_UI_HIPRECY
+#elif DGUS_UI_IS(MKS)
+  #define DGUS_LCD_UI_MKS
+#elif DGUS_UI_IS(RELOADED)
+  #define DGUS_LCD_UI_RELOADED
+#elif DGUS_UI_IS(IA_CREALITY)
+  #define DGUS_LCD_UI_IA_CREALITY
+#endif
+
 /**
  * General Flags that may be set below by specific LCDs
  *
@@ -462,12 +477,11 @@
 #endif
 
 // Aliases for LCD features
-#if ANY(DGUS_LCD_UI_ORIGIN, DGUS_LCD_UI_FYSETC, DGUS_LCD_UI_HIPRECY, DGUS_LCD_UI_MKS)
-  #define HAS_DGUS_LCD_CLASSIC 1
-#endif
-
-#if ANY(HAS_DGUS_LCD_CLASSIC, DGUS_LCD_UI_RELOADED, DGUS_LCD_UI_IA_CREALITY)
+#if !DGUS_UI_IS(NONE)
   #define HAS_DGUS_LCD 1
+  #if DGUS_UI_IS(ORIGIN, FYSETC, HIPRECY, MKS)
+    #define HAS_DGUS_LCD_CLASSIC 1
+  #endif
 #endif
 
 // Extensible UI serial touch screens. (See src/lcd/extui)
