+++ Marlin/src/lcd/extui/mks_ui/tft_multi_language.cpp
@@ -167,7 +167,7 @@ void machine_setting_disp() {
     machine_menu.Extrude_Min_Temper   = EXTRUD_MIN_TEMPER_CN;
 
     machine_menu.HotbedConfTitle      = HOTBED_CONF_TITLE_CN;
-    machine_menu.HotbedAjustType      = HOTBED_ADJUST_CN;
+    machine_menu.HotbedAdjustType     = HOTBED_ADJUST_CN;
     machine_menu.HotbedMinTemperature = HOTBED_MIN_TEMPERATURE_CN;
     machine_menu.HotbedMaxTemperature = HOTBED_MAX_TEMPERATURE_CN;
 
@@ -395,7 +395,7 @@ void machine_setting_disp() {
     machine_menu.Extrude_Min_Temper   = EXTRUD_MIN_TEMPER_T_CN;
 
     machine_menu.HotbedConfTitle      = HOTBED_CONF_TITLE_T_CN;
-    machine_menu.HotbedAjustType      = HOTBED_ADJUST_T_CN;
+    machine_menu.HotbedAdjustType     = HOTBED_ADJUST_T_CN;
     machine_menu.HotbedMinTemperature = HOTBED_MIN_TEMPERATURE_T_CN;
     machine_menu.HotbedMaxTemperature = HOTBED_MAX_TEMPERATURE_T_CN;
 
@@ -627,7 +627,7 @@ void machine_setting_disp() {
 
     machine_menu.HotbedEnable         = HOTBED_ENABLE_EN;
     machine_menu.HotbedConfTitle      = HOTBED_CONF_TITLE_EN;
-    machine_menu.HotbedAjustType      = HOTBED_ADJUST_EN;
+    machine_menu.HotbedAdjustType     = HOTBED_ADJUST_EN;
     machine_menu.HotbedMinTemperature = HOTBED_MIN_TEMPERATURE_EN;
     machine_menu.HotbedMaxTemperature = HOTBED_MAX_TEMPERATURE_EN;
