+++ ini/features.ini
@@ -88,7 +88,7 @@ DGUS_LCD_UI_FYSETC                     = src_filter=+<src/lcd/extui/dgus/fysetc>
 DGUS_LCD_UI_HIPRECY                    = src_filter=+<src/lcd/extui/dgus/hiprecy>
 DGUS_LCD_UI_MKS                        = src_filter=+<src/lcd/extui/dgus/mks>
 DGUS_LCD_UI_ORIGIN                     = src_filter=+<src/lcd/extui/dgus/origin>
-DGUS_LCD_UI_IA_CREALITY                = src_filter=+<src/lcd/extui/Creality>
+DGUS_LCD_UI_IA_CREALITY                = src_filter=+<src/lcd/extui/ia_creality>
 EXTUI_EXAMPLE                          = src_filter=+<src/lcd/extui/example>
 TOUCH_UI_FTDI_EVE                      = src_filter=+<src/lcd/extui/ftdi_eve_touch_ui>
 MALYAN_LCD                             = src_filter=+<src/lcd/extui/malyan>
