+++ platformio.ini
@@ -89,7 +89,7 @@ default_src_filter = +<src/*> -<src/config> -<src/HAL> +<src/HAL/shared> -<src/t
   -<src/lcd/extui/malyan>
   -<src/lcd/extui/mks_ui>
   -<src/lcd/extui/nextion>
-  -<src/lcd/extui/Creality>
+  -<src/lcd/extui/ia_creality>
   -<src/lcd/lcdprint.cpp>
   -<src/lcd/touch/touch_buttons.cpp>
   -<src/sd/usb_flashdrive/lib-uhs2> -<src/sd/usb_flashdrive/lib-uhs3>
